Rotor new rewrite ( again :] )
The previous version of Rotor was, in truth, very verbose and bloaty.

This new version, while still using generational indexes and storages, it's
very different, but especially much more simpler (the examples have half
the LoC compared to the previous version), lightweight and thus faster.

The main change is how the entities components are processed and stored.

Now, entities are really just records with component fields (with an
compile-time addition, just to identify the record as an entity).

The components are literally records identified as components.

There is no more storages for each component, or ids to get these
components, if you have the entity, then the have the components.

These entities live on storages, which stores these entities on a buffer
(actually an array); this makes it less dynamic, but this limitation
doesn't impact development so much.

Now, the systems works directly with entities by processing these storages,
and by analyzing the entity type, the system can decide at compile-time
if all the entities of such type should be ignored or processed, in
other words, the filter step is now compile-time, and not run-time anymore.

There is also no need to mark entities as invalid after it's destruction,
this is done at run-time while running the system, but unlike before,
it's cheap.

Check this commit diff and examples to see how the new version
of Rotor should be used.