System module refactor to better support entity trees and other features
The previous and simpler `run` it's still accepted and supported, and it
will be not deprecated.

Now, systems functions can have a 3rd parameter, which should be an
entity type or concept, it also accepts entities directly instead of
storages on the `run` method.

Many other internal utilities were added in order to better support
entity trees.